Make string primitives PRIMITIVE_UNKNOWN again
Changed the _string_compare, string_concat, string_length, ascii, and
string_contains primitives back to PRIMITIVE_UNKNOWN. Added the method
CallExpr::isPrimitive(const char*) to make it easier to identify primitives
that are PRIMITIVE_UNKNOWN, though it also works with the other primitiveTags.

This removes the need to special case these in the both the parallel pass and
the codegen pass, and cleans up the folding code.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13878 3a8e244f-b0f2-452b-bcba-4c88e055c3ca