minor improvements to optimization flags
Renamed

  --no-expand-iterators-inline-opt
  --no-single-loop-iterator-opt

to

  --no-inline-iterators
  --no-optimize-single-loop-iterators

respectively.

Added

  --no-dead-code-elimination
  --no-live-analysis

Previously these were only available when specifying --baseline or by
using a developer flag respectively.

Made all optimization flags into "n"-style flags so that they can be
disabled (as before) but also enabled, say after using --baseline.

Made --baseline into an aggregate flag that sets the other
optimization flags and is not used throughout the compiler.  Before
this change, it was often used in conjunction with another flag to
ensure that either was set.

Noticed that the description for "N"- and "n"-style flags discusses
the flag as if it were being used one way and this changes from flag
to flag.  For example, the flags

  --[no-]local   Constrain execution to one locale

and

  --[no-]nil-checks   Disable nil checking

are described such that the description matches opposite use cases.
The description is for --local and --no-nil-checks.  How should these
flags be handled?  I followed the lead for the optimization flags
describing one direction.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13882 3a8e244f-b0f2-452b-bcba-4c88e055c3ca