Add a test of nested tuples.
This test nests a tuple 4 deep. It causes the compiler to take a very long
time if inlining is enabled. Adding one more level to the tuple 6*(5*...)
causes a segfault during inlining, but it takes long enough to happen that
the test would time out first in our nightly testing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13889 3a8e244f-b0f2-452b-bcba-4c88e055c3ca