fixed multiple iterator usage bug
Fixed a bug in which an iterator class was iterated over multiple
times at the same time resulting in incorrect behavior when not
inlining iterators.

Inserted code to copy every iterator class into a new iterator class
when starting to iterate.  Since iterator classes are allocated on the
heap, this adversely effects memory usage by storing an extra object
on the heap.  A future optimization to change iterator classes into
records and work with references to these records could fix this
issue.

Added a simple test that shows how an iterator class can be iterated
over multiple times at the same time.  This situation also arises in
op= statements where the left-hand side is an iterator class.

This checkin fixes five of the baseline regressions (all but sparse).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13890 3a8e244f-b0f2-452b-bcba-4c88e055c3ca