fixed baseline + dead-code-elimination regressions
Made it so that deadExpressionElimination is run after
deadCodeElimination and deadVariableElimination to weed out primitive
expressions that have no effect and which we have difficulty
codegening as a result.

Also, disabled one invocation of deadVariableElimination even if
--no-dead-code-elimination is specified.  Made it so that
deadExpressionElimination is run in any event.

This fixes the '--baseline --dead-code-elimination' optimization test
regressions except for sparse.chpl, the only current --baseline
failure.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13891 3a8e244f-b0f2-452b-bcba-4c88e055c3ca