Filing --main-module feature request futures
When I was working on the associative domains/arrays
I found myself surprised that I had to specify a
--main-module flag for a module containing only
inert (non-executable) definitions.  Thinking back,
I remembered that we did this because top-level
variables and constants could be considered
executable and therefore part of a main function,
but for code containing only top-level enums, types,
functions (and possible records, classes, and unions
assuming we never support static constructors), this
seems unnecessarily awkward and unintuitive to me.
Filing some futures to capture these cases.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13893 3a8e244f-b0f2-452b-bcba-4c88e055c3ca