Two pretty-printing futures
These are two futures that I've had in mind since working
on the associative domain/array stuff.  The first requests
that arrays of arrays use linefeeds intuitively (which
might suggest introducing linefeeds after 1D arrays which
I'd also be open to).  The second requests some sort of
mechanism to get a pretty-printed class/record, as in gdb
when throwing a config param (or variable?), e.g.  See
the .future files for more details.

Also, removed the passing valgrind tests from the
REGRESSIONS file.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13894 3a8e244f-b0f2-452b-bcba-4c88e055c3ca