Removed caching of source tree from nightly script
For some reason, the part of this script that
tucked away a copy of the sources each night
has been generating NFS mount errors since I 
moved my run of it to chevy, so I'm removing it.
If we want to continue to do the caching, it
should probably be a separate script anyway.
That said, I've never used these caches, and since
setting it up have become much more confident in
UW and Cray's backup systems, so probably will not
put in the effort to re-enable it.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13898 3a8e244f-b0f2-452b-bcba-4c88e055c3ca