fixed use-before-def bugs with new iterator class copy
Added initializations to fields in iterator classes that are not
references.  Disabled copying of fields that are references in
_getIterator.  (This function is used to shallow copy an iterator
class before iterating as of r13890.)

Added a check to unifyClassInstances in scalarReplace to handle
unifying nil with a class instance since gNil has no definition.

This fixes 3 out of the 4 optimization regressions for the test
release/examples/hpcc/ra with --no-inline-iterators still failing.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13901 3a8e244f-b0f2-452b-bcba-4c88e055c3ca