Took pass over help messages for [no-] flags
Based on Steve's suggestion to use square brackets
in the help line to indicate what the [no-] form
of the flag does, took a pass over the help text
as alluded to in my mail of last night.  This
probably isn't perfect yet, but hopefully is an
improvement.

Also made the booleans set by --fast and --baseline
be static since they are more like meta-flags than
things that should be relied on themselves (they
could probably be removed except that I believe
they're necessary for our --help-settings flag to
work correctly.

Steve: note that I did not take this opportunity
to shorten your very long iterator optimization
flag, though I'd still be happy to see it get
shorter.  :)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13902 3a8e244f-b0f2-452b-bcba-4c88e055c3ca