refactored iterator field initialization for records
Refactored the code we add to the IR to initialize record fields when
an iterator is being created.  Previously, we would take a reference
to the uninitialized record field and initialize the fields in it.
Now, we initialize the fields in a temporary record and then copy that
to the uninitialized record field.  The same recursive strategy is
taken to handle records of records.  This produces much simpler code
and fixes several use-before-def errors seen when not inlining
iterators but applying other optimizations.

This fixes the --no-inline-iterators optimization regression for the
test release/examples/hpcc/ra.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13903 3a8e244f-b0f2-452b-bcba-4c88e055c3ca