improved debug back-tracing through inlining
Added copyMap, a map from BaseAST* to BaseAST* to make it easy to
determine from which BaseAST* another one was copied during inling.
This is very useful if you find yourself in a situation where you are
examining a CallExpr, say, in a large function and you want to
determine what function it was in before inlining happened.  To do
this, simply trace it back through this map (see example in comments)
to find the original.

I added this after having to trace back through inlining twice.
Normally when I debug, I just disable inlining.  But recently trying
to debug some use-before-def errors in the generated code, I couldn't
disable inlining because if I did that, the C compiler would not
report any use-before-def errors.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13904 3a8e244f-b0f2-452b-bcba-4c88e055c3ca