Add a test that causes a deadlock with cobegins.
This test deadlocks with four cobegins that each call two no-op functions
nested in another cobegin. It includes a timeout function so it will run for
at most 30 seconds instead of waiting for the full testing timeout. Also, I
expected that this test would use at most 15 threads, but instead it issues
the warning about not being able to create more threads and there are over
400 threads.

The 15 threads I expect are:
  1 - the main thread.
+ 2 - the begins in the timeout function.
+ 4 - the four inner cobegins
+ 8 - the threads for foo() and bar() within the inner cobegins


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13909 3a8e244f-b0f2-452b-bcba-4c88e055c3ca