Added a warning when a cobegin has fewer than two statements
In order to print the correct line number in the warning, the grammar had to be
changed to save the line number where the cobegin keyword appears, and then this
line number is assigned to the block statement associated with the cobegin.
Technically, the line number of the block statement should be the line where the
left curly brace appears, but the block statement did not consistently have a
relevant line number associated with it (it was often simply 1, regardless of
what line number the block statemnt was on), so it's unlikely that assigning it
the line number where the associated cobegin keyword appears will break
anything.

Removed the .future file associated with the test case I checked in yesterday,
as this change fixes that bug.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13915 3a8e244f-b0f2-452b-bcba-4c88e055c3ca