Add a test that potentially shows a race
This test uses a module from within a begin statement in a loop. Currently it
has the potential to run the initialization function multiple times, though in
practice I've never actually seen it do this without adding a sleep to the
generated C. If it fails, it should print "side effect!" multiple times and
die trying to assign a single variable multiple times.

This race could be fixed by making the "run_module_firsttime" variables into
sync bools, initialized to (full, true). Then every time it is read it is set
to false.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13916 3a8e244f-b0f2-452b-bcba-4c88e055c3ca