Add reporting of blocked threads
Add a new runtime flag --blockreport (-b). When this flag is on, keep track of
threads that are potentially blocked on sync or single variables and trap
SIGINT signals.

Whenever a thread is created a new linked list node is added to a list that
tracks possibly blocked threads. Each thread knows which list node it is
associated with, and when it enters the functions to handle blocking for sync
and single variables, it sets filename and line number information in its node,
as well as a flag to indicate it might be blocked. The SIGINT handler iterates
over this list, and for any node that might be blocked it prints the filename
and line number.

This is only implemented in the pthreads layer, but the flag is available in
all thread layers. I don't know how I can make the flag only be available
where it is functional.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13918 3a8e244f-b0f2-452b-bcba-4c88e055c3ca