Adding sanity checks to malloc calls
Based on user's experience getting segfaults when trying to
allocate more memory than can be expressed by a 32-bit size_t,
put some sanity checks into our memory allocation routines
to catch cases in which the amount of memory requested
exceeds a size_t's capabilities.  Added a test of this that
will only be tested on 32-bit linux machines for now (since
it generates different behavior on 64-bit machines where size_t
is 64 bits).

For the user's particular problem, the allocation was for
2**29 elements of size 2**3 which resulted in a size_t value
of 0.  This was a little harder to track down than I'd
expected because the malloc() implementation on our 32-bit
linux machines returns a non-0x0 pointer for mallocs of
size 0.  As a result, I put in an internal error if chpl_malloc
is called with a size of 0 to see whether we ever do this --
it seems reasonable not to.

It turns out that there were two places that we do -- in
allocating ddata classes for empty arrays and in allocating
internal class instances that have no fields.  It was
reassuring that there weren't more cases that tried to
allocate 0 bytes.

The first case was easy to fix because all array allocations go
through a macro _ARRAY_INIT() which can easily be modified to
guard against the 0 length case and return a 0x0 pointer in that
case.  (In making this change, I also renamed the _ARRAY_INIT
macro and primitive into _ARRAY_ALLOC because the macro/primitive
doesn't actually do any initialization, just the allocation.

The second case was when building classes to pass arguments
to functions that implement 'begin' or 'on' statements.  In
cases where no arguments are being passed in, these classes
are empty and therefore result in mallocs of 0 bytes.  Initially
this case looked easy to fix by simply looking at the number of
arguments and stubbing out the malloc/free calls if there were
none.  In testing, that turned out to be insufficient however,
because filename and lineno fields can be added to these classes
downstream if those values need to be passed into the begin/on
clauses -- and detecting that this will be the case is virtually
impossible.  So instead, I created a new primitive
PRIMITIVE_CHPL_ALLOC_PERMIT_ZERO which is implemented via a macro
that checks for element sizes of 0 and returns 0x0 if that occurs.
This is the one place in the code that uses this primitive and
it would be nice to remove its use in the future.  One potential
way to do this would be to change these classes into records as
Steve has suggested elsewhere.  I didn't take on that change
at this point.






git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13921 3a8e244f-b0f2-452b-bcba-4c88e055c3ca