Fixing setting of CHPL_* variables
My change last week to set CHPL_TARGET_PLATFORM was
invalid because it assumed that CHPL_HOST_PLATFORM
was set and generated an error if it was not.  Fixed
this bug by setting CHPL_TARGET_PLATFORM using the
util script and also set all of the other CHPL_*
variables, if they are not already set, for use by
skipif files.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13924 3a8e244f-b0f2-452b-bcba-4c88e055c3ca