added code to initialize the default wrapper's 'this' symbol
Added PRIMITIVE_INIT_FIELDS which is used to mark the position that
this initialization should be inserted.  This primitive is inserted
during function resolution when default wrappers are created and then
removed at the end of the pass when the types of fields are known.

This should remove the final intel-specific use-before-def warning.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13925 3a8e244f-b0f2-452b-bcba-4c88e055c3ca