Changing CHPL_HOME logic
Previously, the Chapel compiler did its best to determine
what CHPL_HOME was if it was unset, though it didn't do
very well.  In particular, if chpl was run with a relative
or absolute path, it would use that path to determine
where CHPL_HOME was if the environment variable wasn't set.
If CHPL_HOME was not set and if no path was used, it would
(incorrectly) assume that CHPL_HOME was "."  We've had a
comment in the compiler forever indicating that we should
really do something like search the user's path in this
case, but never got around to it.

This hasn't ever been much of an issue because (a) most
people have CHPL_HOME set, and (b) we don't actually rely
on the value of CHPL_HOME much.

My change of last week did rely on CHPL_HOME, and in many
cases, fell into the problem of it being incorrectly
assumed to be "."  This wasn't picked up by the testing
system because the testing system always uses an absolute
path to chpl by default.

In this change, I change the above logic so that CHPL_HOME
will never try to be inferred from the binary's path, in
order to avoid confusion about whether or not CHPL_HOME
needs to be set.  For now, CHPL_HOME needs to be set -- if
it is not, the compiler complains and exits.

I also removed the chplhome flag which, as explained in
my previous check-in, could cause problems due to our
use of CHPL_HOME prior to the parsing of the command-line
flags.  From what I can tell, nobody ever used this much,
and my fear is that it's more confusing than helpful.
If we find a need for it, we can figure out a way to
reintroduce it without causing problems for early needs
of CHPL_HOME.

As far as enabling logic to guess CHPL_HOME if it isn't
unset using a combinatation of argv[0]'s path and the
user's path -- I'm not opposed to this, but was just too
lazy to implement it (as I have been the past several
years apparently).





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13926 3a8e244f-b0f2-452b-bcba-4c88e055c3ca