fixed a couple of bugs in the threading runtime
The first fix in this file handles the case where a thread wakes up
from a wait on the wakeup_signal.  Previously we weren't checking the
task_pool_head pointer.  This change also moves where running_cnt is
decremented and incremented but that should have no effect as it is
still in the critical section.

The second fix replaces the schedule_next_task call to handle the
double-signal case with a pthread_cond_signal call.  This is important
because otherwise we can over-increment waking_cnt (multiple times for
the same task due to repeated signals).

This should fix

  parallel/cobegin/diten/cobeginRace.chpl

on linux.  Waiting to remove the .future based on what Samuel sees on
the mac.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13930 3a8e244f-b0f2-452b-bcba-4c88e055c3ca