Working around free(0x0) with --mem* flags
With my malloc() calls that avoid mallocs of size 0,
frees of 0x0 pointers can cause problems when used
with the --memtrack, --memtrace, etc. flags because
it reports that nothing was ever allocated at 0x0.
The clean way to deal with this would be to never
call _chpl_free() on 0x0 and to introduce a primitive
FREE_PERMIT_NULL to mirror ALLOC_PERMIT_ZERO, but I
didn't have time to implement that today (or better,
remove the need for ALLOC_PERMIT_ZERO), so am
checking this in as a workaround -- basically just
a conditional to avoid any of the memtracking code
firing when 0x0 is freed.

Worth noting: tried changing the class used by begin
and on clauses into a record rather than a class as
alluded to in my previous checkin on this topic, and
found that a class can't be used in the begin case
because it needs to outlive the stack on which it
would be allocated.  However, the begin case never
uses an empty class, so it turned out that making
the on clause use a record and the begin case a class
worked.  Which is to say, the tests all passed.

Then Steve pointed out that with ARMCI and segment
fast, the other downside of using records is that
they can't refer to stack-allocated variables,
shooting that whole exploration down in an instant.

So, next approach I'd like to try is to use 
primitives like ALLOC_PERMIT_ZERO and FREE_PERMIT_NULL,
and to have the compiler remove these or convert them
into normal ALLOC/FREE primitives depending on whether
they have any fields or not after the insertLinenos pass.

Also worth noting: if making CHPL_FREE throw an internal
error on 0x0, the same tests fail as were failing due
to allocations of size 0:

[Error matching program output for arrays/deitz/jacobi-no-local]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/defaultValueGlobalTuple]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/fieldDefaultGlobalArray]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/fieldDefaultGlobalClassMember]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/fieldDefaultGlobalRecordMember]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/fieldDefaultGlobalTuple]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/recordFieldDefaultGlobalTuple]
[Error matching program output for
multilocale/diten/nolocalArgDefaultGlobal/simpleDefaultValueGlobal]




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13933 3a8e244f-b0f2-452b-bcba-4c88e055c3ca