fixed iterator bug regarding fields in iterator class
Since r10860, compiling with --no-live-analysis resulted in iterator
classes being defined by adding every local variable in the iterator
(not just the variables that are live at yield points).  Before this
change, local references were omitted.  This became a problem when
loop indices could capture references.  Since these references could
actually be live at yield points, they need to be added to the
iterator class as fields.  Thus r10860 removed this omission.

This checkin adds this omission back in so that references are not
added to the iterator class.  However, index variables are added even
if they are references.  (Index variables are now marked with the
pragma "index var".)  This is necessary because blindly adding all
references makes it impossible for cullOverReferences to remove
references that need to be removed in order that the implicit 'setter'
argument is correctly set.

This checkin also fixes an inefficiency in which the returned symbol
could be added as a field twice.

This should fix the baseline and optimization regressions related to
the sparse test case.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13935 3a8e244f-b0f2-452b-bcba-4c88e055c3ca