Fixing logic in compiler.pl script
Recent changes to the start_test script (setting 
CHPL_* environment variables) and compiler
(querying CHPL_* settings through scripts) revealed 
a bug in the compiler.pl script in which the host
compiler would be taken as the target compiler
if the CHPL_HOST_COMPILER variable is set.
This should only occur if the variable is set
and the CHPL_HOST_PLATFORM equals the
CHPL_TARGET_PLATFORM (e.g., I should be able
to set CHPL_HOST_COMPILER to intel and for a
non-cross-compile on linux64, have intel used
for both the front-end and back-end compiler;
yet if I am compiling for the MTA or XT and
have chosen gcc as my CHPL_HOST_COMPILER, I
shouldn't also get that as my CHPL_TARGET_COMPILER).

Updated the script to fix this bug and updated
the README.chplenv to clarify the relationship
between these variables.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13936 3a8e244f-b0f2-452b-bcba-4c88e055c3ca