removed --[no-]flow-analysis
This flag was an aggregate flag that (more or less) enabled [disabled]
live-analysis, dead-code-elimination, and copy-propagation, all
optimizations that require flow analyses.  However, it was not
implemented as an aggregate and was set with baseline.  So for example
doing

  --baseline --live-analysis

did not actually reenable live analysis.  I could have fixed this by
making the --[no-]flow-analysis flag into an aggregate so that it set
the other flags.  I decided that it is not all that useful in general,
however, so decided to remove it.

Also change

  --[no-]optimize-single-loop-iterators

to

  --[no-]optimize-loop-iterators


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13938 3a8e244f-b0f2-452b-bcba-4c88e055c3ca