Default args in extern function test/future
I've been doing some extern function work, and happily
discovered that it's possible to declare default
arguments in the Chapel definition of the external
function and have Chapel build the wrapper that makes
this work as one would expect.  Nice!

I also found a case that doesn't work: for an extern
function that has a default argument and returns a
value, the return type is lost in the wrapper,
presumably because there is no function body to
analyze.  This seems like it should be possible to
support, so I filed a feature request future for it.
If it isn't possible/reasonable to support then we
should specialize the error message for this case.

This is low priority, but if it's easy to fix would
be an attractive feature to use (in particular, the
rewrite of the memory tests that I'm working on
could make use of them).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13949 3a8e244f-b0f2-452b-bcba-4c88e055c3ca