Adding opaque types
Adding support for opaque types, for use in interfacing
with external C routines that require types which cannot
be expressed easily in Chapel to be passed/returned.  As
an example, if you were calling a C function that
returned a pointer to a struct that you wanted to pass
into a second function, you could use an opaque type to
represent that pointer within the Chapel code.  Opaque
types are not intended to support more than argument
passing and assignment for like types.

Implemented this by introducing new dtOpaque and gOpaque
datatypes/symbols into the compiler and adding a special
case to the parser's type rules to handle opaque since
it's a keyword.  Based on discussions with Steve, will
deprecate opaque from keyword to standard identifier,
much like our other types (int, bool, etc.) in a distinct
checkin.

Added some tests that use opaque types in simple ways
using inference, assignment, explicit declarations,
and type aliases.

Added a description of opaque types to the README.extern
file.  Chose not to add it to the language spec at this
point since its role is primarily in external interfaces
at this point, and those are not documented in the spec
(or even officially part of the language).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13951 3a8e244f-b0f2-452b-bcba-4c88e055c3ca