Chapel-ized Shannon's memory tests
Rewrote Shannon's memory tests in Chapel using our extern
capabilities to call directly into the functions in
chplmem.c that they rely on.  In doing so, removed the
chplmem_tests.c file from the runtime (yay!) and the
calls that it exported from the standard modules (wahoo!)
and wrote the logic that they contained in Chapel.

The other big benefit of this change is that I rewrote
tests that relied on classes being a certain size to
simply use pure malloc's in order to increase their
portability.  The tests were designed more to test our
memory tracking system than to lock in the size of a
particular class, so this is kosher and will make them
far more portable.  As a result, I hope/expect that
we won't have any more platform-specific memory failures
in tomorrow's runs (or at least, soon...).

Unrelated, but caught up in this checkin: updated
cygwin regressions based on yesterday's changes.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13952 3a8e244f-b0f2-452b-bcba-4c88e055c3ca