specialized the parameter tuple indexing method
Added code to function resolution preFold to catch a call of the
parameter tuple indexing function before it is resolved.  It is
replaced by a member access of the appropriate field in the tuple,
thus avoiding the need to have a tuple indexing function.

Removed code used to special-case the instantiation of the tuple
indexing method.

Fixed a bug in cullOverReferences in which a reference could be
incorrectly culled if a field of it were accessed and stored in a
temporary that itself had to be a reference.  In short,

  y = &(x.y); (*y).z = ...

could become

  y = x.y; y.z = ...

This never showed up before because the usage of the GET_MEMBER
primitive never introduced such code.  With the replacement of
variable functions by GET_MEMBER primitives directly, such code can
now be generated.

Removed some dead code in cullOverReferences.

This change makes an error message more specific; updated a .good.

On chevy, this cuts the time to compile

  test/types/tuple/diten/tupleOfTupleOfTupleOfTuple.chpl

from 54.7 seconds to 25.2s, a speedup of 54%.  This is an improvement
but there is more to do.  For jacobi, the time is cut from 8.40s to
7.97s.

Comparing the compiler before and after this change and with compiling
the compiler with debug or with using OPTIMIZE=1, we see (in seconds)

               debug compiler           optimized compiler
               before   after           before  after
tupleOfTu...   54.7     25.2  (54%)     42.1    19.2  (54%)
jacobi          8.40     7.97 (5.1%)     6.53    6.28 (3.8%)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13953 3a8e244f-b0f2-452b-bcba-4c88e055c3ca