applied reference propagation to pre-inlined functions
During the inlining pass, we optimize functions that we are about to
inline in order to keep the intermediate representation smaller.  This
change folds singleAssignmentRefPropagation into the mix before
localCopyPropagation, much like is done later in the optimization
passes.

The further decreases the time to compile

  test/types/tuple/diten/tupleOfTupleOfTupleOfTuple.chpl

from 25.2s to 19.0s, a speedup of nearly 25%.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13960 3a8e244f-b0f2-452b-bcba-4c88e055c3ca