optimized verify (done between passes)
Moved the code to check back pointers (parentExpr and parentSymbol)
into the dynamically dispatch verify methods.  Prior to this change,
these back pointers were recomputed in a traversal over the entire
AST.  Removed this traversal in favor of running over gAsts, the
global vector of AST nodes.  Along with decreasing the time spent in
verification significantly (by about a factor of six), this has the
advantage that verification is computed in a substantively different
way than how the back pointers were originally set.

Moved the code in passes/verify.cpp to AST/baseAST.cpp as the code in
this file is now only a few lines and this is a more consistent place
for it (as the other code to clean up the IR between passes is here).

Although this decreases the time spent between passes significantly
(this time includes cleanAST as well as verify), the total compilation
time is improved only slightly.  It appears that the time spent in
each pass is increased slightly as the time spent between passes is
decreased.  I've noticed this before when I did experiments where I
just removed verification.  My guess is that this has to do with
caching at some level as the order visited by the traversal may be
more optimal when moving onto a new pass than the order visited by
looping over gAsts.

For jacobi, I'm seeing (in seconds)

                      before       after
time between passes   1.11         0.58
total time            7.68         7.54


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13963 3a8e244f-b0f2-452b-bcba-4c88e055c3ca