Remove block-level-statement
Remove the block-level-statement concept from the language. This was used to
make certain statements be allowed anywhere a block-statement is allowed, for
example as the only statement in a for loop without the 'do' keyword. Now,
if a for loop has only one statement it has to be in a block-statement or the
'do' keyword is always needed.

Similar changes were made everywhere a 'do' or 'then' keyword shows up, such
as 'when' statements, 'on' statements, and 'if' statements.

Made sync statements create a new scope, even for the one-statement version.

Added a one-statement version of the serial statement which previously only
allowed a block-level-statement, so something like "serial true do foo();"
wasn't allowed.

Functions and methods, which previously allowed any block-level-statement as
the function body can now only have a block statement, empty statement or
return statement as the body.

Updated the spec and tests to reflect these changes.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13967 3a8e244f-b0f2-452b-bcba-4c88e055c3ca