Don't allow empty statements in functions without blocks
Now only return statements are allowed to be used for functions that don't use
braces to open blocks.

test_where_param2.chpl

Was only passing because the semicolons in the where clauses were taken as the
body of the functions. The blocks with the writelns would be executed even if
the functions were never called. If this was the intent of this test, I'm not
sure how this behavior could be reproduced.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13971 3a8e244f-b0f2-452b-bcba-4c88e055c3ca