refactored compute_sym_uses
Background: The compute_sym_uses(...) routine builds up the vectors
stored in each Symbol as Symbol::uses and Symbol::defs.  The uses
vector contains every SymExpr that is a use of the symbol.  The defs
vector contains every SymExpr that is a def of the symbol.

Implemented the following variants of this routine:

  compute_sym_uses()
  builds the vectors for every variable/argument in the entire program

  compute_sym_uses(FnSymbol* fn)
  builds the vectors for every variable/argument in 'fn' and looks for
  uses and defs only in 'fn'

  compute_sym_uses(Vec<Symbol*>& set)
  builds the vectors for every variable/argument in 'set' and looks
  for uses and defs in the entire program; the vectors should be
  cleared before calling this variant

  compute_sym_uses(Vec<Symbol*>& set, Vec<BaseAST*>& asts)
  builds the vectors for every variable/argument in 'set' and looks
  for uses and defs only in 'asts'; the vectors should be cleared
  before calling this variant

I plan to refactor these once more soon.

Note that currently in heapAllocateGlobals and heapAllocateLocals, we
use the last variant even though we look for uses and defs in the
entire program; this is because the defs vector needs to be built up
in an ordered way such that the first def is the def where the heap
allocation should happen.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13978 3a8e244f-b0f2-452b-bcba-4c88e055c3ca