Re-enabling evaluation of on clauses for side effects
In r13480, Steve made it so that using the --local flag does
not evalute the locality expression argument of an on clause.
This resulted in the strange behavior that when running with
CHPL_COMM = none on a single locale, things like:

          on Locales(1)

would not result in out-of-bounds expressions.  This struck
me as being surprising since there aren't two locales.
Taking a look, we saw that when the --serial flag is used,
serial expressions are still evaluated for side effects,
and that the apparent motivating case for not evaluating
the on clauses could be handled in other ways.  

In particular, when compiling --local I now wrap the on
clause expression in the same primitive that is used in
the --no-local case to handle them in the same manner.
This seems to cause everything to work, and generates
a nice out-of-bounds error for the on clause in the
above example.

Updated the man page to note this behavior as we do for
the --serial flag.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13982 3a8e244f-b0f2-452b-bcba-4c88e055c3ca