changed data structure used in the symbol table
Made SymScope::table into a Map instead of a ChainHashMap.  The
ChainHashMap leaks memory and also uses significantly more than Map.
This does require all strings in the table to be canonicalized so that
identical strings have the same address.  Since we already do this,
this change was straightforward.

Canonicalized the --main-module string before looking up the symbol as
well and also canonicalized several strings passed to lookup wrapper
functions.

For arrays/deitz/jacobi, this decreases the leaked memory from
2,121,120 bytes in 130,293 blocks to 1,014,903 bytes in 38,127 blocks.
Note these numbers are consistent but not deterministic.  This is our
biggest leak.  I plan to close a few other sizeable leaks as well.

I'm also seeing about a 6% speedup in compiling this test on chevy.
Interestingly, this speedup is not accounted for by the scope
resolution pass which shows only a nominal difference.  It is
accounted for in the other passes, in particular, flattenFunctions and
some of the optimizations.  I've mentioned before that I'd like to see
us build and use a symbol table only during the scope resolution pass
and thus not need to maintain it throughout compilation.  This is
non-trivial work (though I don't think it will be particularly
difficult).  This provides some evidence that it may be worthwhile.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13986 3a8e244f-b0f2-452b-bcba-4c88e055c3ca