Making test system print relative paths
(This is a change that I polled about when we added support
for start_test to be run from arbitrary paths, but have
not been inconvenienced enough to add until now).

This change causes start_test to report test names in
terms of a relative path from the directory in which
start_test was executed, rather than relative to the
root $CHPL_HOME/test directory.  The motivation for
this is so that if a test fails, one can execute a
start_test command giving the same test name as listed
in the summary message without manually adjusting for
the fact that they may not be in the test/ directory.

As an example, if you invoked start_test from the
extern/bradc directory and had errors in that directory
and in the opaque/ subdirectory, you would see output like:

[Error matching program output for ./defaultArgs]
[Error matching program output for opaque/opaqueAlias]

Note that if you run from the $CHPL_HOME directory, this
means that test/ will be added to the test names that
you're used to getting.  It seems like this should be
similarly helpful, but if you don't like this, let
me know -- it'd be easy enough to special-case.

It's possible that this change will cause some of our
cron jobs to report diffs that don't really exist
because of the differences in the test paths, but I've
manually checked all the cases I can think of and
believe that they shouldn't cause problems -- in most
cases, they were written to execute from the test/
directory (start_opt_tests may not be, but also doesn't
compute diffs from one night to the next, so won't
generate a bad report due to this change).

Let me know if you have any problems, of course.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@13989 3a8e244f-b0f2-452b-bcba-4c88e055c3ca